requirements.txt
langchain==0.0.138
nltk==3.7
openai==0.27.2
pinecone_client==2.2.1
python-dotenv==1.0.0
PyYAML==6.0
beautifulsoup4==4.12.2